refactor: Extract constant to another file
