+++ app/components/header-constants.ts
@@ -1,0 +1,1 @@
+export const HEADER_HEIGHT = 92;
